# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"VisitorResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"cpf": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"cpf",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VisitResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"visitor": {
						"$ref": "#/components/schemas/VisitorResponse"
					},
					"date": {
						"type": "string"
					},
					"residentId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"visitor",
					"date",
					"residentId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VisitListResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"residentId": {
						"type": "string"
					},
					"visits": {
						"items": {
							"$ref": "#/components/schemas/VisitResponse"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"residentId",
					"visits",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VisitaCreateRequest": {
				"properties": {
					"visitante": {
						"properties": {
							"descricao": {
								"type": "string"
							},
							"tipo": {
								"type": "string"
							},
							"cpf": {
								"type": "string"
							},
							"nome": {
								"type": "string"
							}
						},
						"required": [
							"cpf",
							"nome"
						],
						"type": "object"
					},
					"data": {
						"type": "string"
					}
				},
				"required": [
					"visitante",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ListaVisitasRequest": {
				"properties": {
					"visitas": {
						"items": {
							"$ref": "#/components/schemas/VisitaCreateRequest"
						},
						"type": "array"
					}
				},
				"required": [
					"visitas"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "TCC API",
		"version": "1.0.0",
		"description": "API para gerenciamento de moradores, visitantes e visitas",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/visitas": {
			"get": {
				"operationId": "GetVisitas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VisitListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Lista de visitas não encontrada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Buscar lista de visitas por ID do morador",
				"tags": [
					"Visitas"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "moradorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOrUpdateVisitas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VisitListResponse"
								}
							}
						}
					},
					"400": {
						"description": "Dados inválidos",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Morador não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Inserir/atualizar lista de visitas",
				"tags": [
					"Visitas"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "moradorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ListaVisitasRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteVisita",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Lista de visitas não encontrada",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Deletar visita da lista",
				"tags": [
					"Visitas"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "moradorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"visitaId": {
										"type": "string"
									}
								},
								"required": [
									"visitaId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "https://localhost:3000/api"
		}
	]
}